apps/api/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic[dotenv]
python-jose[cryptography]
httpx
psycopg[binary]
